Removed kitchen package and related code
This commit removes the usage of the kitchen package to handle the single
function of converting file data to bytes. As the functionality and properties
of the file data is unclear (should it be text data or can it be other?), I
choose to use the generic built-in bytes type instead of the string method
encode.

In addition to get a base level check that these would be equivalent I made
a small, and I understand incomplete, unit test - tests/unit/kitchen_bytes.py
to confirm some level of equivalency. I don't think after a couple releases
this unit test needs to stick around.

Also this change does not look at the current state of Push File keyword. In
reviewing the reason for to_bytes, it looks as if the functionality of driver's
push_filer method has changes since the opriginal addition of this keyword. I
suggest another issue to review this. This commit only removes kitchen.